+++ src/events/base_events.py
@@ -9,6 +9,7 @@ from typing import TYPE_CHECKING
 
 import ops
 
+from common.custom_events import RestartWorkloadEvent, UnitFullyStartedEvent
 from common.exceptions import (
     RequestingLockTimedOutError,
     ValkeyACLLoadError,
@@ -41,54 +42,10 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class RestartWorkloadEvent(ops.EventBase):
-    """Event for restarting the workload when certain events happen, e.g. IP change."""
-
-    def __init__(
-        self, handle: ops.Handle, restart_valkey: bool = True, restart_sentinel: bool = True
-    ):
-        super().__init__(handle)
-        self.restart_valkey = restart_valkey
-        self.restart_sentinel = restart_sentinel
-
-    def snapshot(self) -> dict[str, str]:
-        """Save the state of the event."""
-        return {
-            "restart_valkey": str(self.restart_valkey),
-            "restart_sentinel": str(self.restart_sentinel),
-        }
-
-    def restore(self, snapshot: dict[str, str]) -> None:
-        """Restore the state of the event."""
-        self.restart_valkey = snapshot.get("restart_valkey", "True") == "True"
-        self.restart_sentinel = snapshot.get("restart_sentinel", "True") == "True"
-
-
-class UnitFullyStarted(ops.EventBase):
-    """Event that signals that the unit's has fully started.
-
-    This event will be deferred until:
-        The Sentinel service is running and was discovered by other units.
-        The Valkey service is running and the current node is in sync with the primary (if a replica).
-    """
-
-    def __init__(self, handle: ops.Handle, is_primary: bool = False):
-        super().__init__(handle)
-        self.is_primary = is_primary
-
-    def snapshot(self) -> dict[str, str]:
-        """Save the state of the event."""
-        return {"is_primary": str(self.is_primary)}
-
-    def restore(self, snapshot: dict[str, str]) -> None:
-        """Restore the state of the event."""
-        self.is_primary = snapshot.get("is_primary", "False") == "True"
-
-
 class BaseEvents(ops.Object):
     """Handle all base events."""
 
-    unit_fully_started = ops.EventSource(UnitFullyStarted)
+    unit_fully_started = ops.EventSource(UnitFullyStartedEvent)
     restart_workload = ops.EventSource(RestartWorkloadEvent)
 
     def __init__(self, charm: "ValkeyCharm"):
@@ -206,7 +163,7 @@ class BaseEvents(ops.Object):
         )
 
     # TODO check how to trigger if deferred without update status event
-    def _on_unit_fully_started(self, event: UnitFullyStarted) -> None:
+    def _on_unit_fully_started(self, event: UnitFullyStartedEvent) -> None:
         """Handle the unit-fully-started event."""
         if not self.charm.cluster_manager.is_healthy(
             is_primary=event.is_primary, check_replica_sync=False
