+++ src/events/external_clients.py
@@ -22,6 +22,7 @@ from dpcharmlibs.interfaces import (
     ValkeyResponseModel,
 )
 
+from common.custom_events import TopologyChangedEvent
 from common.exceptions import (
     KubernetesClientError,
     ValkeyACLLoadError,
@@ -41,16 +42,6 @@ if TYPE_CHECKING:
     from charm import ValkeyCharm
 
 logger = logging.getLogger(__name__)
-
-
-class TopologyChangedEvent(ops.EventBase):
-    """A custom event for topology changes."""
-
-
-class TopologyChangedCharmEvents(ops.CharmEvents):
-    """A CharmEvent extension to observe topology changes."""
-
-    topology_changed = ops.EventSource(TopologyChangedEvent)
 
 
 class ExternalClientsEvents(ops.Object):
